pyproject.toml
[project]
name = "xxdm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "schedule>=1.2.2",
]
